refactor: unbraid Hickey findings — extract pipeline, highlight, fix line map
1. Extract renderPlanMarkdown into planMarkdown.ts as a pipeline of
   independent transforms (stampLineNumbers, restyleFeedback) — each
   can be tested/reordered independently.

2. Extract change highlighting into usePlanChangeHighlight.ts — own
   module, own concern, no coupling to rendering or selection.

3. Fix buildLineMap first-occurrence-wins: now stores ALL occurrences
   per text snippet with a consumption index, so duplicate lines
   (e.g. repeated table values) get sequential line numbers.

PlanPane.tsx drops from 464 to 270 lines, composing 3 modules.
